fix:integer to double 로 cost 타입 수정

+++ backend/src/main/java/com/almostThere/domain/map/entity/link/MapLink.java
@@ -4,11 +4,11 @@ import lombok.Getter;
 
 @Getter
 public class MapLink {
-    private final int map_start_id;
-    private final int map_end_id;
-    private final int cost;
+    private final long map_start_id;
+    private final long map_end_id;
+    private final double cost;
 
-    public MapLink(int map_start_id, int map_end_id, int cost) {
+    public MapLink(long map_start_id, long map_end_id, double cost) {
         this.map_start_id = map_start_id;
         this.map_end_id = map_end_id;
         this.cost = cost;
